Editing cabins

+++ src/services/apiCabins.js
@@ -1,27 +1,37 @@
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
-  
-  const { data, error } = await supabase
-    .from('cabins')
-    .select('*')
+  const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.log(error);
-    throw new Error('Cabins could not be loaded')
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
   }
 
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    ""
+  );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -29,19 +39,24 @@ export async function createCabin(newCabin) {
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
-    .from('cabin-images')
+    .from("cabin-images")
     .upload(imageName, newCabin.image);
-  
-  // 3. Delete the cabin IF there was an error uploading the image
+
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
-    throw new Error("Cabin image could not be uploaded and the cabin was not created");
-  } 
-  
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
+    );
+  }
+
   return data;
 }
+
 export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
